Remove unnecessary casts.

+++ src/Compressed.cpp
@@ -41,7 +41,7 @@ OutputInfo Compressed::initialize(std::vector<double*>& parameterPointers,
 
   // Bias component will not change after initialization
   if(bias)
-    y(J) = double(1.0);
+    y(J) = 1.0;
 
   initializeParameters();
